Add user selection by optional regex
The default behavior is retained when passing -u without an additional
pattern. Alternatively, a regex pattern can be parsed that must match
exactly one group, that is the user-id to login. An example pattern is
'^(?:user|mail|id): (.*)'

Fixes handling of files with only one line when -u overrides the
(default) -P option by copying a space into the selection.